+++ src/exactly_lib/impls/types/string_source/defs.py
@@ -16,9 +16,10 @@ SOURCE_FILE_ARGUMENT_NAME = a.Named('SOURCE-FILE-PATH')
 
 
 def src_rel_opt_arg_conf_for_phase(phase_is_after_act: bool,
+                                   default_relativity: RelOptionType = RelOptionType.REL_CWD,
                                    ) -> RelOptionArgumentConfiguration:
     return source_file_relativities.src_rel_opt_arg_conf_for_phase(
-        RelOptionType.REL_CWD,
+        default_relativity,
         SOURCE_FILE_ARGUMENT_NAME.name,
         phase_is_after_act,
     )