+++ CHANGELOG.md
@@ -44,7 +44,8 @@ with exception of a "0." prefix until version 1 is released.
  - Exit codes - Syntax error in `[act]` - `VALIDATION_ERROR` -> `SYNTAX_ERROR`
  - Exit codes - Values of all non-zero exit codes
  - Instruction `conf`/`actor` - Remove "-" prefix from actor names
- - Builtin symbol - string `LINE_SEP` -> Rename to `OS_LINE_SEP` 
+ - Instruction `setup`/`stdin` - Contents value is `STRING-SOURCE`
+ - Builtin symbol - string `LINE_SEP` -> Rename to `OS_LINE_SEP`
 
 ### Removed